SourceCode
In our project we put together a neural network that could make a binary choice, whether it is a Volkswagen Passat or a Fiat Doblo, given the car photos. This neural network, when trained for more car brands and models, can be used in security cameras placed in the roads and streets to collect data about how owners of each car model react to different traffic situations, whether they prefer to take a longer road with less traffic or if they would choose to go directly with more traffic, and help manufacturers tweak their cars according to driver preferences.